+++ query_expansion/llm_based_code/qa_expand.sh
@@ -12,11 +12,11 @@
 
 module load cuda/11.8
 source /work/pi_wenlongzhao_umass_edu/27/.venv/bin/activate
-MODEL="deepseek"
-mkdir -p results/cosqa_queries_expanded_${MODEL}_temp_0.5_tfidf results/cosqa_queries_expanded_${MODEL}_temp_0.5_mmr logs
+MODEL="granite"
+mkdir -p results/cosqa/cosqa_queries_expanded_${MODEL}_temp_0.5_tfidf2 results/cosqa/cosqa_queries_expanded_${MODEL}_temp_0.5_mmr2 logs
 INPUT_CSV="data/cosqa_queries/part_${SLURM_ARRAY_TASK_ID}.csv"
-OUTPUT_CSV1="results/cosqa_queries_expanded_${MODEL}_temp_0.5_tfidf/result_${SLURM_ARRAY_TASK_ID}.csv"
-OUTPUT_CSV2="results/cosqa_queries_expanded_${MODEL}_temp_0.5_mmr/result_${SLURM_ARRAY_TASK_ID}.csv"
+OUTPUT_CSV1="results/cosqa/cosqa_queries_expanded_${MODEL}_temp_0.5_tfidf2/result_${SLURM_ARRAY_TASK_ID}.csv"
+OUTPUT_CSV2="results/cosqa/cosqa_queries_expanded_${MODEL}_temp_0.5_mmr2/result_${SLURM_ARRAY_TASK_ID}.csv"
 
 
 echo "Running task ${SLURM_ARRAY_TASK_ID} on host $(hostname)"
@@ -27,4 +27,4 @@ if [ ! -f "$INPUT_CSV" ]; then
     exit 1
 fi
 
-python query_exp3.py $MODEL $INPUT_CSV $OUTPUT_CSV1 $OUTPUT_CSV2
+python queryexp_main.py $MODEL $INPUT_CSV $OUTPUT_CSV1 $OUTPUT_CSV2
